# Improve performance of notify-references