--- a/lib/src/clojure_lsp/feature/diagnostics.clj
+++ b/lib/src/clojure_lsp/feature/diagnostics.clj
@@ -175,8 +175,8 @@
 (defn ^:private file-var-definitions [project-analysis filename]
   (q/find-var-definitions project-analysis filename false))
 (def ^:private file-kw-definitions q/find-keyword-definitions)
-(def ^:private project-var-definitions q/find-all-var-definitions)
-(def ^:private project-kw-definitions q/find-all-keyword-definitions)
+(def ^:private all-var-definitions q/find-all-var-definitions)
+(def ^:private all-kw-definitions q/find-all-keyword-definitions)
 
 (defn custom-lint-project!
   [new-analysis kondo-ctx db]
@@ -188,6 +188,16 @@
       (lint-defs! (all-var-definitions project-analysis)
                   (all-kw-definitions project-analysis)
                   project-analysis kondo-ctx))))
+
+(defn custom-lint-files!
+  [filenames new-analysis kondo-ctx db]
+  (let [project-analysis (into {}
+                               (q/filter-project-analysis-xf db)
+                               new-analysis)
+        file-analyses (select-keys project-analysis filenames)]
+    (lint-defs! (all-var-definitions file-analyses)
+                (all-kw-definitions file-analyses)
+                project-analysis kondo-ctx)))
 
 (defn custom-lint-file!
   [filename analysis kondo-ctx db]