--- a/lib/src/clojure_lsp/feature/file_management.clj
+++ b/lib/src/clojure_lsp/feature/file_management.clj
@@ -33,7 +33,7 @@
                       (update-analysis uri (:analysis kondo-result))
                       (update-findings uri (:findings kondo-result))
                       (assoc :kondo-config (:config kondo-result)))))
-      (f.diagnostic/async-lint-file! uri db)))
+      (f.diagnostic/async-publish-diagnostics! uri db)))
   (when-let [new-ns (and allow-create-ns
                          (string/blank? text)
                          (contains? #{:clj :cljs :cljc} (shared/uri->file-type uri))
@@ -114,7 +114,7 @@
           (logger/debug "Analyzing references for files:" filenames)
           (crawler/analyze-reference-filenames! filenames db)
           (doseq [filename filenames]
-            (f.diagnostic/sync-lint-file! (shared/filename->uri filename db) db))
+            (f.diagnostic/sync-publish-diagnostics! (shared/filename->uri filename db) db))
           (producer/refresh-code-lens producer))))))
 
 (defn ^:private offsets [lines line col end-line end-col]
@@ -174,7 +174,7 @@
                                                   (update :processing-changes disj uri)
                                                   (assoc :kondo-config (:config kondo-result))))
               (do
-                (f.diagnostic/sync-lint-file! uri db)
+                (f.diagnostic/sync-publish-diagnostics! uri db)
                 (when (settings/get db [:notify-references-on-file-change] true)
                   (notify-references filename old-local-analysis (get-in @db [:analysis filename]) components))
                 (clojure-producer/refresh-test-tree producer [uri]))
@@ -206,7 +206,7 @@
                       (update :analysis merge analysis)
                       (assoc :kondo-config (:config result))
                       (update :findings merge (group-by :filename (:findings result))))))
-      (f.diagnostic/lint-project-files! filenames db)
+      (f.diagnostic/publish-all-diagnostics! filenames db)
       (clojure-producer/refresh-test-tree producer uris))))
 
 (defn did-change-watched-files [changes db]
@@ -238,7 +238,7 @@
                                      (shared/dissoc-in [:documents uri])
                                      (shared/dissoc-in [:analysis filename])
                                      (shared/dissoc-in [:findings filename]))))
-        (f.diagnostic/clean! uri db)))))
+        (f.diagnostic/publish-empty-diagnostics! uri db)))))
 
 (defn force-get-document-text
   "Get document text from db, if document not found, tries to open the document"