--- a/lib/src/clojure_lsp/feature/file_management.clj
+++ b/lib/src/clojure_lsp/feature/file_management.clj
@@ -10,6 +10,7 @@
    [clojure-lsp.shared :as shared]
    [clojure.core.async :as async]
    [clojure.java.io :as io]
+   [clojure.set :as set]
    [clojure.string :as string]
    [lsp4clj.protocols.logger :as logger]
    [lsp4clj.protocols.producer :as producer]
@@ -84,26 +85,31 @@
     (find-changed-elems-by usage-signature old-var-usages new-var-usages)))
 
 (defn reference-filenames [filename old-local-analysis new-local-analysis db]
-  (let [project-analysis (into {} (q/filter-project-analysis-xf db) (:analysis @db))
-        source-paths (settings/get db [:source-paths])
-        changed-var-definitions (find-changed-var-definitions old-local-analysis new-local-analysis)
-        references-filenames (->> changed-var-definitions
-                                  ;; TODO: switch from nested-loop join to hash join
-                                  (mapcat #(q/find-references project-analysis % false db))
-                                  (map :filename))
+  (let [changed-var-definitions (find-changed-var-definitions old-local-analysis new-local-analysis)
         changed-var-usages (find-changed-var-usages old-local-analysis new-local-analysis)
-        definitions-filenames (->> changed-var-usages
-                                   ;; TODO: remove usages of external namespaces (clojure.core, etc.) here
-                                   ;; TODO: switch from nested-loop join to hash join
-                                   (keep #(q/find-definition project-analysis % db))
-                                   (filter (fn [d]
-                                             ;; TODO: this excludes "private calls", but they are counted in code lens, so maybe shouldn't exclude
-                                             (and (not (:private d))
-                                                  ;; TODO: is this extra work? The definition came from project-analysis, so should be on the source-paths.
-                                                  (some #(string/starts-with? (:filename d) %) source-paths))))
-                                   (map :filename))]
-    (disj (set (concat references-filenames definitions-filenames))
-          filename)))
+        project-analysis (into {}
+                               (q/filter-project-analysis-xf db)
+                               (dissoc (:analysis @db) filename)) ;; don't notify self
+        source-paths (settings/get db [:source-paths])
+        references-filenames (into #{}
+                                   (comp
+                                     ;; TODO: switch from nested-loop join to hash join
+                                     (mapcat #(q/find-references project-analysis % false db))
+                                     (map :filename))
+                                   changed-var-definitions)
+        definitions-filenames (into #{}
+                                    (comp
+                                      ;; TODO: remove usages of external namespaces (clojure.core, etc.) here
+                                      ;; TODO: switch from nested-loop join to hash join
+                                      (keep #(q/find-definition project-analysis % db))
+                                      (filter (fn [d]
+                                                ;; TODO: this excludes "private calls", but they are counted in code lens, so maybe shouldn't exclude
+                                                (and (not (:private d))
+                                                     ;; TODO: is this extra work? The definition came from project-analysis, so should be on the source-paths.
+                                                     (some #(string/starts-with? (:filename d) %) source-paths))))
+                                      (map :filename))
+                                    changed-var-usages)]
+    (set/union references-filenames definitions-filenames)))
 
 (defn ^:private notify-references [filename old-local-analysis new-local-analysis {:keys [db producer]}]
   (async/go