--- a/lib/src/clojure_lsp/feature/file_management.clj
+++ b/lib/src/clojure_lsp/feature/file_management.clj
@@ -87,29 +87,44 @@
 (defn reference-filenames [filename old-local-analysis new-local-analysis db]
   (let [changed-var-definitions (find-changed-var-definitions old-local-analysis new-local-analysis)
         changed-var-usages (find-changed-var-usages old-local-analysis new-local-analysis)
+        usage-signature (juxt :to :name)
+        def-signatures (fn [var-def]
+                         (map (fn [var-name]
+                                [(:ns var-def) var-name])
+                              (q/var-definition-names var-def)))
         project-analysis (into {}
                                (q/filter-project-analysis-xf db)
                                (dissoc (:analysis @db) filename)) ;; don't notify self
-        source-paths (settings/get db [:source-paths])
-        references-filenames (into #{}
-                                   (comp
-                                     ;; TODO: switch from nested-loop join to hash join
-                                     (mapcat #(q/find-references project-analysis % false db))
-                                     (map :filename))
-                                   changed-var-definitions)
-        definitions-filenames (into #{}
-                                    (comp
-                                      ;; TODO: remove usages of external namespaces (clojure.core, etc.) here
-                                      ;; TODO: switch from nested-loop join to hash join
-                                      (keep #(q/find-definition project-analysis % db))
-                                      (filter (fn [d]
-                                                ;; TODO: this excludes "private calls", but they are counted in code lens, so maybe shouldn't exclude
-                                                (and (not (:private d))
-                                                     ;; TODO: is this extra work? The definition came from project-analysis, so should be on the source-paths.
-                                                     (some #(string/starts-with? (:filename d) %) source-paths))))
-                                      (map :filename))
-                                    changed-var-usages)]
-    (set/union references-filenames definitions-filenames)))
+        references-filenames (when (seq changed-var-definitions)
+                               (let [def-signs (into #{}
+                                                     (mapcat def-signatures)
+                                                     changed-var-definitions)]
+                                 (into #{}
+                                       (comp
+                                         (mapcat val)
+                                         (filter #(identical? :var-usages (:bucket %)))
+                                         (filter #(contains? def-signs (usage-signature %)))
+                                         (map :filename))
+                                       project-analysis)))
+        definitions-filenames (when (seq changed-var-usages)
+                                (let [usage-signs->langs (->> changed-var-usages
+                                                              ;; TODO: can we easily remove usages of external namespaces (clojure.core, etc.) here?
+                                                              ;; TODO: do we really care if lang doesn't match?
+                                                              (reduce (fn [result usage]
+                                                                        (assoc result (usage-signature usage) (q/elem-langs usage)))
+                                                                      {}))]
+                                  (into #{}
+                                        (comp
+                                          (mapcat val)
+                                          (filter #(identical? :var-definitions (:bucket %)))
+                                          (filter #(when-let [usage-langs (some usage-signs->langs (def-signatures %))]
+                                                     (some usage-langs (q/elem-langs %))))
+                                          ;; TODO: this excludes "private calls", but they are counted in code lens, so maybe shouldn't exclude
+                                          (remove :private)
+                                          (map :filename))
+                                        project-analysis)))]
+    (set/union (or references-filenames #{})
+               (or definitions-filenames #{}))))
 
 (defn ^:private notify-references [filename old-local-analysis new-local-analysis {:keys [db producer]}]
   (async/go