… lib/src/clojure_lsp/queries.clj
… lib/test/clojure_lsp/features/diagnostics_test.clj