This replaces the nested-loop joins in notify-references with hash joins.

This PR doesn't really change the performance when there aren't any files to notify. But when there are, it makes a difference.

To test I inserted an unmatched delimiter into `clojure-lsp.feature.diagnostics`, making the file invalid. This file uses 5 namespaces, and 5 other namespaces use it. They all get notified.

Before this PR, in this scenario notify-references takes about 2 seconds. About half of that is finding the files and the other half is analyzing and linting them. After the patch it takes about 1 second, with about 100ms finding files, 800ms analyzing and 100ms linting. 

For follow-on work, finding files and linting each execute `q/filter-project-analysis`, which takes about 100ms itself, meaning that almost all their time is in that function. So, making `q/filter-project-analysis` faster will improve `notify-references` further.

In addition to removing the nested loops, this PR:

* Renames diagnostic helpers to delineate which calculate custom lint data and which publish that data to clients.
* Lints all reference files in one batch, rather than linting each file individually

- [x] This is follow up work from #885

This PR builds on #887. You can either merge that one first and this second, or do them both by merging this one alone.