@ericdallo 2022-04-02
All working!
One thing I noticed, it seems kondo doesn't lint vars when used as `(#'my-alias/foo ...)`, not sure if this is on purpose or should be fixed on kondo, but our lint of private calls are useless, right?

Edit:

Not really, just found that if we remove the function, that private call using `#'` gets a `unresolved-var` warning, so it's still valid to lint private calls